chore: moved package structure to more conventional locations, optimized imports
