+++ backend/services/event-service/src/main/java/com/ticket4u/Application.java
@@ -1,4 +1,4 @@
-package com;
+package com.ticket4u;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
