+++ backend/services/user-service/src/main/java/com/ticket4u/service/AuthService.java
@@ -1,11 +1,11 @@
 package com.ticket4u.service;
 
+import com.nimbusds.jose.JOSEException;
 import com.ticket4u.constant.TokenConstants;
 import com.ticket4u.dto.auth.AuthResponse;
 import com.ticket4u.dto.auth.LoginRequest;
 import com.ticket4u.dto.auth.internal.LoginResult;
 import com.ticket4u.entity.CustomUserDetails;
-import com.nimbusds.jose.JOSEException;
 import com.ticket4u.util.CookieUtil;
 import com.ticket4u.util.JwtUtil;
 import com.ticket4u.util.TokenUtil;
